Bug 1727172 - Use doubles instead of ints when computing video scale for fallback painting. r=jrmuizel
Differential Revision: https://phabricator.services.mozilla.com/D123411

UltraBlame original commit: 4e0dafb873bd3b45c0cecf4a8e64f7e324211f7c

+++ layout/generic/nsVideoFrame.cpp
@@ -635,8 +635,8 @@ class nsDisplayVideo : public nsPaintedDisplayItem {
                         static_cast<int32_t>(destGFXRect.Height()));
     
     SwapScaleWidthHeightForRotation(scaleToSize, rotationDeg);
-    transform.PreScale(scaleToSize.width / size.Width(),
-                       scaleToSize.height / size.Height());
+    transform.PreScale(scaleToSize.width / double(size.Width()),
+                       scaleToSize.height / double(size.Height()));
 
     gfxContextMatrixAutoSaveRestore saveMatrix(aCtx);
     aCtx->SetMatrix(
